a quick hack to make sever run in Nuke10.x
Added a test for version and call object directly if nuke version is above 9.x. 
Not sure what (if any) issues this might cause, but sees to behave in my tool using this package.

I've filed a bug report at The Foundry regarding calls to nuke.executeInMainThreadWithResult() not working the same way in Nuke10.x.